доработка дз

+++ hw5/hw5.py
@@ -1,5 +1,6 @@
 import datetime
-import os
+import os.path
+import requests
 
 
 class MediaData:
@@ -22,44 +23,91 @@ class MediaData:
     def __repr__(self):#для преобразованя в читаемую форму
         return f'{self.value}'
 
-    def __add__(self, other: int):#для работы add как сумма чисел
-        return MediaData(self.value + other.value)
+    def __add__(self, other):#для работы add как сумма чисел
+        return MediaData(self.file_size + other)
+
 
 class Video(MediaData):
     # здесь будут методы типа save, open, rename, delete в зависимости от типов файлов
+    def download_from_url(self, url, save_path):
+        pass
 
-    def read_(self):
-        # Здесь будет код для чтения из облака
-        print(f"Чтение файла {self.path}")
-        return f'чтение файла {self.path}'
+    def upload_to_s3(self, file_path, bucket_name, object_name=None):
+        pass
 
     pass
 
 class Audio(MediaData):
     # здесь будут методы типа save, open, rename, delete в зависимости от типов файлов
-    def delite(self):
-        return os.remove(self.name)
+    def delite(self, path):
+        if os.path.exists(path):
+            print(f'файл {path} удалён')
+            return os.remove(self.path)
+        else:
+            print(f'файла не существует')
     pass
 
 class Photo(MediaData):
     # здесь будут методы типа save, open, rename, delete в зависимости от типов файлов
 
-    def rename(self, new_name):
-        return os.rename(self.name, new_name)
+    def rename(self,path, new_name):
+        if os.path.exists(path):
+            print(f'файл {path} переименован')
+            return os.rename(self.name, new_name)
+        else:
+            print(f'файла не существует')
+
     pass
 
-print(os.getcwd())
-
 #объявляем переменные
-media3 = Video('C:/Users/Инна/Desktop/WIN_20150711_124528.JPG', 4, 'df', (2025, 2, 4), 'C:/Users/Инна/Desktop/WIN_20150711_124528.JPG')
-mf_path = 'C:/Users/Инна/Desktop/Dimasik_.mp3'
-media1 = Photo('C:/Users/Инна/Desktop/WIN_20150711_124528.JPG',2,'inna', datetime.date(2024, 12, 2), mf_path)
-photo_path = 'C:/Users/Инна/Desktop/Dima.mp3'
-media2 = Audio('C:/Users/Инна/Desktop/Dima.mp3', 3,'nata', datetime.date(2025, 1, 3), photo_path)
-video1 = Video('https://docs.google.com/spreadsheets/d/1vqtrZbLnc_mQqKTMTN4dJJ38pcwoLKscdAfWhsZBQ3s/edit?usp=sharing', 5, 'inna_b', (2024, 2, 12), 'https://docs.google.com/spreadsheets/d/1vqtrZbLnc_mQqKTMTN4dJJ38pcwoLKscdAfWhsZBQ3s/edit?usp=sharing')
+media3 = Video('C:/Users/Инна/Desktop/WIN_20150711_124528.JPG', 4, 'df', datetime.date(2020, 11,14), 'C:/Users/Инна/Desktop/WIN_20150711_124528.JPG')
+mf_path = 'C:/Users/Инна/Desktop/IMG_20200630_195058.jpg'
+media1 = Photo(mf_path,2,'inna', datetime.date(2024, 12, 2), mf_path)
+photo_path = 'C:/Users/Инна/Desktop/IMG_20200630_195128.jpg'
+media2 = Audio(photo_path, 3,'nata', datetime.date(2025, 1, 3), photo_path)
+
+
 
 #пробуем работают ли методы
-print((media1.file_size + media2.file_size))
-media1.rename('C:/Users/Инна/Desktop/my_photo.JPG')#переименование файла
-media2.delite()
-video1.read_()
+# print((media1.file_size + media2.file_size))
+# media1.rename(media1.path,'C:/Users/Инна/Desktop/my_photo2.JPG')#переименование файла
+# media2.delite(media2.path) #проверка удаления файла
+# print(media3.creation_date)
+
+
+class CloudStorageMixin(MediaData):
+# """Mixin для работы с файлами в облаке и удаленных хранилищах."""
+
+    def download_from_url(self, url, save_path):
+        response = requests.get(url, stream=True)
+        if response.status_code == 200:
+            with open(save_path, 'wb') as file:
+                for chunk in response.iter_content(1024):
+                    file.write(chunk)
+                    print(f'Файл скачан: {save_path}')
+        else:
+            print(f'Ошибка скачивания: {response.status_code}')
+
+    def upload_to_s3(self, file_path, bucket_name, object_name=None):
+        s3 = boto3.client('s3')
+        if object_name is None:
+            object_name = os.path.basename(file_path)
+            try:
+                s3.upload_file(file_path, bucket_name, object_name)
+                print(f'Файл {file_path} загружен в S3 как {object_name}')
+            except NoCredentialsError:
+                print('Ошибка: Не указаны учетные данные для AWS S3')
+
+# В облаке:
+video_cloud = Video(
+'https://example.com/video.mp4', # URL удаленного файла
+10, 'cloud_user', datetime.date(2024, 2, 24),
+'/tmp/video.mp4' # Временный путь для скачивания
+)
+video_cloud.download_from_url(video_cloud.name, video_cloud.path) # Скачивание
+video_cloud.upload_to_s3(video_cloud.name,"my-bucket") # Загрузка в S3
+
+
+
+
+
